refactor(bundle-size): extract createRootDummyManager behind getRootDummyInputs
The dummy-input factory and its dependencies (createDummyInputManager,
DummyInputManagerCore, focusedElement.getFirstOrLastTabbable,
keyboardNavigation.setNavigatingWithKeyboard, nativeFocus) accounted for
~3 kB inside Root.js + DummyInput.js even when an app didn't want
browser-Tab handling. Move the factory to a new src/RootDummyManager.ts
and gate Root.addDummyInputs() on tabsterCore.rootDummyManagerFactory,
which is registered by a new public getRootDummyInputs(tabster) helper.

Without that opt-in, Root.addDummyInputs() is a no-op even when
controlTab/rootDummyInputs is true. Apps that need browser-Tab handling
call getRootDummyInputs(tabster) once after createTabster (and after the
flag is set) — same shape as the other getX factories.

The internal RootAPI.onRoot handshake is gone (folded into createRoot
and _onRootDispose by an earlier commit), so the factory only needs the
core, the WeakHTMLElement, the setFocused callback, and sys.

Also wire test-setup.js and .storybook/preview.mjs to call
getRootDummyInputs when controlTab/rootDummyInputs is set so existing
tests/stories continue to behave the same.

createTabster: 31.65 kB → 28.54 kB (-3.11 kB / -9.8%)
DummyInput.js:  5.54 kB → 2.73 kB (-2.81 kB)
Root.js:        2.88 kB → 2.60 kB (-280 B)

Breaking change: controlTab=true (default) no longer auto-creates
dummy inputs; consumers that need browser Tab handling must now call
getRootDummyInputs(tabster) once.
